Boost.Variant customization into a separate header
