+++ code/xtl.cpp
@@ -47,117 +47,9 @@
 #include "boost/variant.hpp"
 #include "type_switchN-patterns-xtl.hpp"
 #include "patterns/all.hpp"
+#include "adapt_boost_variant.hpp"
 
 //----------------------------------------------------------------------------------------------------------------------
-
-#define VARIANT_P(params) BOOST_PP_ENUM_PARAMS(10, params)
-
-// NOTE: We declare vtbl_of in boost namespace because we want it to be found 
-//       via two-phase name lookup due to its argument boost::variant
-namespace boost
-{
-#if XTL_SUPPORT(variadics)
-    // Default implementation of vtbl_of grabs sizeof(intptr_t) bytes from the beginning
-    // of the object, which on 64-bit machines is larger that size of variant's which member.
-    template <class... Ts>
-    inline std::intptr_t vtbl_of(const boost::variant<Ts...>* p) { return p->which(); }
-#else
-    template <VARIANT_P(class Ts)>
-    inline std::intptr_t vtbl_of(const boost::variant<VARIANT_P(Ts)>* p) { return p->which(); }
-#endif
-}
-
-namespace xtl
-{
-#if XTL_SUPPORT(variadics)
-    template <class... Ts>
-    struct is_poly_morphic<boost::variant<Ts...>>
-    {
-        static const bool value = true;
-    };
-
-    template <class S, class T, class... Ts>
-    struct is_subtype<S,boost::variant<T,Ts...>>
-    {
-        static const bool value = is_subtype<S,T>::value
-                               || is_subtype<S,boost::variant<Ts...>>::value;
-    };
-#else
-    template <VARIANT_P(class Ts)>
-    struct is_poly_morphic<boost::variant<VARIANT_P(Ts)>>
-    {
-        static const bool value = true;
-    };
-
-    template <class S, class T, VARIANT_P(class Ts)>
-    struct is_subtype<S,boost::variant<T,VARIANT_P(Ts)>>
-    {
-        static const bool value = is_subtype<S,T>::value
-                               || is_subtype<S,boost::variant<VARIANT_P(Ts)>>::value;
-    };
-#endif
-
-    template <class S, class T>
-    struct is_subtype<S,boost::variant<T>> : is_subtype<S,T> {};
-
-#if XTL_SUPPORT(variadics)
-	template <class... Ts, class S>
-    boost::variant<Ts...> subtype_cast_impl(target<boost::variant<Ts...>>, const S& s)
-    {
-        return boost::variant<Ts...>(s); // FIX: Actually this should be boost::variant<Ts...>(xtl::subtype_cast<Ti>(s)) where S <: Ti
-    }
-#else
-    template <VARIANT_P(class Ts), class S>
-    boost::variant<VARIANT_P(Ts)> subtype_cast_impl(target<boost::variant<VARIANT_P(Ts)>>, const S& s)
-    {
-        return boost::variant<VARIANT_P(Ts)>(s); // FIX: Actually this should be boost::variant<Ts...>(xtl::subtype_cast<Ti>(s)) where S <: Ti
-    }
-#endif
-
-    template <typename T>
-    struct is_subtype_visitor : public boost::static_visitor<T*>
-    {
-        template <typename S>
-        T* operator()(S& s) const
-        {
-            return xtl::subtype_dynamic_cast<T*>(&s);
-        }
-    };
-
-#if XTL_SUPPORT(variadics)
-    template <class T, class... Ts>
-    typename std::enable_if<xtl::is_subtype<T, boost::variant<Ts...>>::value, T*>::type
-    subtype_dynamic_cast_impl(target<T*>, boost::variant<Ts...>* pv)
-    {
-        is_subtype_visitor<T> visitor;
-        return boost::apply_visitor(visitor, *pv);
-    }
-
-    template <class T, class... Ts>
-    typename std::enable_if<xtl::is_subtype<T, boost::variant<Ts...>>::value, const T*>::type
-    subtype_dynamic_cast_impl(target<const T*>, const boost::variant<Ts...>* pv)
-    {
-        is_subtype_visitor<const T> visitor;
-        return boost::apply_visitor(visitor, *pv);
-    }
-#else
-    template <class T, VARIANT_P(class Ts)>
-    typename std::enable_if<xtl::is_subtype<T, boost::variant<VARIANT_P(Ts)>>::value, T*>::type
-    subtype_dynamic_cast_impl(target<T*>, boost::variant<VARIANT_P(Ts)>* pv)
-    {
-        is_subtype_visitor<T> visitor;
-        return boost::apply_visitor(visitor, *pv);
-    }
-
-    template <class T, VARIANT_P(class Ts)>
-    typename std::enable_if<xtl::is_subtype<T, boost::variant<VARIANT_P(Ts)>>::value, const T*>::type
-    subtype_dynamic_cast_impl(target<const T*>, const boost::variant<VARIANT_P(Ts)>* pv)
-    {
-        is_subtype_visitor<const T> visitor;
-        return boost::apply_visitor(visitor, *pv);
-    }
-#endif
-}
 
 static_assert(xtl::is_subtype<int,int>::value,"No reflexivity");
